agentic_socratic_nlp_tutor/pyproject.toml
[project]
name = "agentic_socratic_nlp_tutor"
version = "0.1.0"
description = "Agentic_Socratic_NLP_Tutor using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.100.0,<1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-chroma>=0.1.0",
    "langchain-huggingface>=0.0.1",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
agentic_socratic_nlp_tutor = "agentic_socratic_nlp_tutor.main:run"
run_crew = "agentic_socratic_nlp_tutor.main:run"
train = "agentic_socratic_nlp_tutor.main:train"
replay = "agentic_socratic_nlp_tutor.main:replay"
test = "agentic_socratic_nlp_tutor.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
